pyproject.toml
[project]
name = "iss-group-24"
version = "0.3.0"
description = "Few-shot multi-shot localizer (OWLv2) + siamese existence (DINOv2)."
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0",
    "Pillow>=10.0",
    "scipy>=1.10",
    "torch>=2.4",
    "torchvision>=0.19",
    "transformers>=4.40",
    "accelerate>=0.30",
    "peft>=0.10",
    "huggingface_hub>=0.23",
    "matplotlib>=3.7",
]
